openclaw-telegram-bot/requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
aiohttp==3.9.1
